Finalized comparePercents.js restAPI to return an array[4] of of percentages of lifters the user is stronger than in: total, bench, squat, and deadlift. Combined these queries into a single statement to reduce sum of total latency from mySQL connection/HTML requests
